+++ docs/release_notes.rst
@@ -5,6 +5,7 @@ Release Notes
 0.9.3
 -----
 
+* Upgraded Node.js to ``v14.17.2`` `(2021-07-01) <https://github.com/nodejs/node/blob/master/doc/changelogs/CHANGELOG_V14.md#14.17.2>`_
 * Added ``IV8ModuleResolver`` to allow intercepting module resolving callback
 
 0.9.2